fix(tether): removed reposition on window scroll
In Cloud Office CP we were experiencing a bug when a popover opened on the
bottom of a page could open in a location below the page that was unscrollable
in the browser. Removing the window scroll event as a trigger for repositioning
the popover fixed it and seems to have no unintended consequences that we
found, but depending on the reason for having it there in the first place,
maybe it does in some condition.