Locate rs-popover-target at click-event-time
The rsPopoverTrigger directive will -- in some circumstances -- get
created and run `link()` before templates like the following get a chance
to be evaluated:

<some-element id="{{key}}"></some-element>

Which means findPopoverTarget won't find it and will throw an error even
though an element with the specified id will eventually exist.

A common case where this may happen is when the rs-popover-target exists
in a repeater and is pointed to an element of the repeater whose id can
only be generated by an angular binding expression.